-Replace hardcoded path with moduleProperty in sql queries. -Remove broken query study.currentBloodAvailable. -Add module property DCM_NHP_Resources_Container. -Update automated tests.
